Rename new files to suit style guide
