+++ drake/solvers/equality_constrained_qp_solver.h
