+++ drake/solvers/linear_system_solver.cc
